+++ src/SegGeometryMatrixV1.py
@@ -357,9 +357,9 @@ class SegGeometryMatrixV1(SegGeometry) :
                  +4 - mask non-bonded pixels
                  +8 - mask nearest neighbours of nonbonded pixels
         """
-        zero_col = np.zeros(sp._rows,dtype=np.int)
-        zero_row = np.zeros(sp._cols,dtype=np.int)
-        mask     = np.ones((sp._rows,sp._cols),dtype=np.int)
+        zero_col = np.zeros(sp._rows,dtype=np.uint8)
+        zero_row = np.zeros(sp._cols,dtype=np.uint8)
+        mask     = np.ones((sp._rows,sp._cols),dtype=np.uint8)
 
         if mbits & 1 : 
         # mask edges
